# Distribute Money Transfer Platform/bank-service/src/main/java/com/azry/dmtp/z/model/api/HomeResponse.java
package com.azry.dmtp.z.model.api;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@Data
@NoArgsConstructor
@AllArgsConstructor
public class HomeResponse {
    private String greeting;
}