# Distribute Money Transfer Platform/dmtp/fi-integration/src/main/resources/openapi/bank-service.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bank API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/bank-service",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "tags": [
    {
      "name": "accounts",
      "description": "The accounts API"
    },
    {
      "name": "clients",
      "description": "The clients API"
    },
    {
      "name": "transactions",
      "description": "The transactions API"
    },
    {
      "name": "home",
      "description": "The home API"
    }
  ],
  "paths": {
    "/api/v1/client/{id}": {
      "get": {
        "tags": [
          "clients"
        ],
        "operationId": "getClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "clients"
        ],
        "operationId": "updateClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/client/deactivate/{id}": {
      "put": {
        "tags": [
          "clients"
        ],
        "operationId": "deactivateClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountNumber}/withdraw": {
      "put": {
        "tags": [
          "accounts"
        ],
        "operationId": "withdraw",
        "parameters": [
          {
            "name": "accountNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amountOfMoney",
            "in": "query",
            "required": true,
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountNumber}/deposit": {
      "put": {
        "tags": [
          "accounts"
        ],
        "operationId": "deposit",
        "parameters": [
          {
            "name": "accountNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amountOfMoney",
            "in": "query",
            "required": true,
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/deactivate/{accountNumber}": {
      "put": {
        "tags": [
          "accounts"
        ],
        "operationId": "deactivateAccount",
        "parameters": [
          {
            "name": "accountNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions": {
      "post": {
        "tags": [
          "transactions"
        ],
        "operationId": "createTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/client": {
      "post": {
        "tags": [
          "clients"
        ],
        "operationId": "createClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCreationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "operationId": "search_2",
        "parameters": [
          {
            "name": "pagingRequest",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagingRequest"
            }
          },
          {
            "name": "accountFilter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AccountFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageAccountDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "operationId": "createAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{id}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "operationId": "getTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/filter": {
      "get": {
        "tags": [
          "transactions"
        ],
        "operationId": "search",
        "parameters": [
          {
            "name": "pagingRequest",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagingRequest"
            }
          },
          {
            "name": "transactionFilter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TransactionFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTransactionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/home": {
      "get": {
        "tags": [
          "home"
        ],
        "operationId": "getGreeting",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HomeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/client/personal-number/{personalNumber}": {
      "get": {
        "tags": [
          "clients"
        ],
        "operationId": "getClientByPersonalNumber",
        "parameters": [
          {
            "name": "personalNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/client/filter": {
      "get": {
        "tags": [
          "clients"
        ],
        "operationId": "search_1",
        "parameters": [
          {
            "name": "pagingRequest",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagingRequest"
            }
          },
          {
            "name": "clientFilter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ClientFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageClientDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountNumber}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "operationId": "getAccount",
        "parameters": [
          {
            "name": "accountNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountNumber}/balance": {
      "get": {
        "tags": [
          "accounts"
        ],
        "operationId": "checkBalance",
        "parameters": [
          {
            "name": "accountNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/client-accounts/{clientId}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "operationId": "getAccounts",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountDTO"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClientUpdateRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "citizenship": {
            "type": "string"
          }
        }
      },
      "AccountDTO": {
        "required": [
          "accountNumber",
          "accountStatus",
          "balance",
          "clientId",
          "currency",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "accountNumber": {
            "type": "string"
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "balance": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "clientId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ClientDTO": {
        "required": [
          "address",
          "citizenship",
          "clientStatus",
          "dateOfBirth",
          "email",
          "firstName",
          "id",
          "lastName",
          "personalNumber",
          "phone"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "personalNumber": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "citizenship": {
            "type": "string"
          },
          "clientStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "accountDTOList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountDTO"
            }
          }
        }
      },
      "BalanceDTO": {
        "required": [
          "accountNumber",
          "balance",
          "currency"
        ],
        "type": "object",
        "properties": {
          "balance": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          }
        }
      },
      "TransactionCreationRequest": {
        "required": [
          "amount",
          "currency",
          "receiverAccountNumber",
          "senderAccountNumber"
        ],
        "type": "object",
        "properties": {
          "receiverAccountNumber": {
            "type": "string"
          },
          "senderAccountNumber": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        }
      },
      "TransactionDTO": {
        "required": [
          "amount",
          "creationDate",
          "currency",
          "id",
          "receiverAccountNumber",
          "senderAccountNumber",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "receiverAccountNumber": {
            "type": "string"
          },
          "senderAccountNumber": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING",
              "PROCESSED",
              "REJECTED"
            ]
          }
        }
      },
      "ClientCreationRequest": {
        "required": [
          "address",
          "citizenship",
          "clientStatus",
          "dateOfBirth",
          "email",
          "firstName",
          "lastName",
          "personalNumber",
          "phone"
        ],
        "type": "object",
        "properties": {
          "personalNumber": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "citizenship": {
            "type": "string"
          },
          "clientStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "AccountCreationRequest": {
        "required": [
          "balance",
          "clientId",
          "currency"
        ],
        "type": "object",
        "properties": {
          "balance": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "clientId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagingRequest": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TransactionFilter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "receiverAccountNumber": {
            "type": "string"
          },
          "senderAccountNumber": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "searchStartDate": {
            "type": "string",
            "format": "date-time"
          },
          "searchEndDate": {
            "type": "string",
            "format": "date-time"
          },
          "amountLowerBound": {
            "type": "number"
          },
          "amountUpperBound": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING",
              "PROCESSED",
              "REJECTED"
            ]
          }
        }
      },
      "PageTransactionDTO": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDTO"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string"
          },
          "nullHandling": {
            "type": "string"
          },
          "ascending": {
            "type": "boolean"
          },
          "property": {
            "type": "string"
          },
          "ignoreCase": {
            "type": "boolean"
          }
        }
      },
      "HomeResponse": {
        "type": "object",
        "properties": {
          "greeting": {
            "type": "string"
          }
        }
      },
      "ClientFilter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "personalNumber": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "citizenship": {
            "type": "string"
          },
          "clientStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "PageClientDTO": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientDTO"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "AccountFilter": {
        "required": [
          "accountNumber",
          "clientId",
          "currency",
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "currency": {
            "type": "string"
          },
          "clientId": {
            "type": "integer",
            "format": "int64"
          },
          "accountNumber": {
            "type": "string"
          }
        }
      },
      "PageAccountDTO": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountDTO"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}